requirements.txt
python-socketio
fastapi
uvicorn
motor
requests
azureml-fsspec
opencv-python